% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_pred_vs_pro.R
\name{add_pred_to_pro}
\alias{add_pred_to_pro}
\title{Add prediction mean to professional network dataset}
\usage{
add_pred_to_pro(pred, pro)
}
\arguments{
\item{pred}{prediction dataset with lat, lon, pred_mean and pred_sd columns}

\item{pro}{professional weather station dataset}
}
\value{
the professional dataset with added columns pred_mean and pred_sd
}
\description{
Add prediction mean to professional network dataset.
This function is needed to run model evaluation.
}
\author{
Eva Marques
}
