% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAMBA_Functions.R
\name{Analyze_Samba_Genes}
\alias{Analyze_Samba_Genes}
\title{Analyze_Samba_Genes}
\usage{
Analyze_Samba_Genes(
  sgRes,
  ntc.as.null.dist = T,
  score.method = "MetaAnalysis",
  file.prefix = NULL
)
}
\arguments{
\item{sgRes}{data.frame of the results from the guide-level screen enrichment analysis. This is the output from the "Analyze_Samba_Guides" function.}

\item{ntc.as.null.dist}{logical, indicating whether to ese NTCs or all guides to generate a null distribution for gene-level analyses.}

\item{score.method}{character, indicating whether to use the 'MetaAnalysis' or 'GeneScore' method to perform gene-level analysis.}

\item{file.prefix}{(optional) file.path and/or character prefix to use for exporting the result tables}
}
\value{
data.frame of the results from the gene-level screen enrichment analysis.
}
\description{
This function performs the gene-level screen enrichment analysis of SAMBA, using guide-level SAMBA results as an input.
}
