% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Samba_functions.R
\name{Analyze_Samba_Guides}
\alias{Analyze_Samba_Guides}
\title{Analyze_Samba_Guides}
\usage{
Analyze_Samba_Guides(
  dge,
  method = "QLF",
  design = NULL,
  coefficient = NULL,
  contrast = NULL,
  GuideMap = NULL,
  file.prefix = NULL,
  verbose = T
)
}
\arguments{
\item{dge}{DGEList object with normalization and dispersion already calculated. This should be the output from the "Preprocess_Samba" function.}

\item{method}{character, indicating which statistical model and test to use for guide-level analysis. Currently, only "QLF" and "LRT" are supported.}

\item{design}{design matrix for the samples in "data". Note: The order of the rows must match that of the data read-counts.}

\item{coefficient}{character name or vector, indicating which coefficient of the linear model is tested to be equal to zero.}

\item{contrast}{contrast matrix, indicating the contrast(s) of the linear model to be tested as equal to zero (optional).}

\item{GuideMap}{data.frame that maps the guides of the library to their respective genes (optional). Must have "sgRNA" and "Gene" columns.}

\item{file.prefix}{(optional) file.path and/or character prefix to use for exporting the result tables}

\item{verbose}{logical, indicating whether to display verbose output information.}
}
\value{
data.frame of the results from the guide-level screen enrichment analysis.
}
\description{
This function performs guide-level screen enrichment analysis of SAMBA, using a pre-processed DGEList object as an input.
}
