% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_pred_vs_pro.R
\name{boxplot_per_hour}
\alias{boxplot_per_hour}
\title{Plot residuals boxplot per hour from reference network}
\usage{
boxplot_per_hour(pro)
}
\arguments{
\item{pro}{The dataset of professional weather stations used as reference}
}
\value{
A ggplot object displaying the residual boxplots
}
\description{
Plot residuals boxplot per hour from reference network
}
