% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{check_area_shp}
\alias{check_area_shp}
\title{Check area shapefile format}
\usage{
check_area_shp(area_shp)
}
\arguments{
\item{area_shp}{area shapefile}
}
\description{
Check area shapefile  format
}
\author{
Eva Marques
}
