% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{check_cws_raw}
\alias{check_cws_raw}
\title{Check raw cws format}
\usage{
check_cws_raw(cws_raw)
}
\arguments{
\item{cws_raw}{citizen weather station dataset}
}
\description{
Check raw cws format
}
\author{
Eva Marques
}
