% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{check_era5_accum}
\alias{check_era5_accum}
\title{Check era5 accumulated format}
\usage{
check_era5_accum(era5_accum)
}
\arguments{
\item{era5_accum}{era5 accumulated}
}
\description{
Check era5 accumulated format
}
\author{
Eva Marques
}
