% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{check_era5_instant}
\alias{check_era5_instant}
\title{Check era5 instantaneous format}
\usage{
check_era5_instant(era5_instant)
}
\arguments{
\item{era5_instant}{era5 instantaneous}
}
\description{
Check era5 instantaneous format
}
\author{
Eva Marques
}
