% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{check_fch}
\alias{check_fch}
\title{Check forest canopy height format}
\usage{
check_fch(fch)
}
\arguments{
\item{fch}{forest canopy height}
}
\description{
Check forest canopy height format
}
\author{
Eva Marques
}
