% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{check_input}
\alias{check_input}
\title{Check input list format}
\usage{
check_input(input)
}
\arguments{
\item{input}{a list with all elements needed for running the Bayesian model}
}
\description{
Check the content of the input list
}
\author{
Eva Marques
}
