% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{check_ts_te}
\alias{check_ts_te}
\title{Check time start and time end format}
\usage{
check_ts_te(ts, te)
}
\arguments{
\item{ts}{time start}

\item{te}{time end}
}
\description{
Check time start and time end  format
}
\author{
Eva Marques
}
