% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_prediction_grid.R
\name{create_empty_grids}
\alias{create_empty_grids}
\title{Create 100km*100km grids inside US cells provided by their id}
\usage{
create_empty_grids(cells_id)
}
\arguments{
\item{cells_id}{a numeric vector of cell ids. See find_cells() for more
details. Use 1:1450 to create grids for the whole US.}
}
\value{
a list of sf objects. The ith element of the list corresponds to the
ith cell id in US 100km resolution grid.
}
\description{
Create 100km*100km grids within US extent
}
