% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_prediction_grid.R
\name{create_st_grids_from_cells}
\alias{create_st_grids_from_cells}
\title{Create monthly 100km*100km grids inside US cells}
\usage{
create_st_grids_from_cells(cells_id, yyyy, mm, directory)
}
\arguments{
\item{cells_id}{a numeric vector of cell ids. See find_cells() for more
details. Use 1:1450 to create grids for the whole US.}

\item{yyyy}{a numeric year}

\item{mm}{a numeric month}

\item{directory}{a character path to the grid folder}
}
\description{
Create 100km*100km grids within US extent
}
