% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_marginals.R
\name{density_beta_covar}
\alias{density_beta_covar}
\title{Plot density for spatial covariate coefficients from the BHM}
\usage{
density_beta_covar(info)
}
\arguments{
\item{info}{the dataframe from BHM output with inference informations}
}
\value{
a ggplot object
}
\description{
Plot density for spatial covariate coefficients from the BHM
}
\author{
Eva Marques
}
