% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_covariates.R
\name{download_covariates}
\alias{download_covariates}
\title{Function to download covariates for a given area and time period}
\usage{
download_covariates(covar_list, area, ts, te, storage_dir)
}
\arguments{
\item{covar_list:}{list of covariates to download. Items must be in
c("elev", "gridmet", "daymet", "prism", "nlcd", "lcz", "pop", "build_fp", "build_h",
"imp", "tcc", "fch", "era5")}

\item{area:}{area of interest (a sf, sfc, SpatRaster or SpatVector object)}

\item{ts:}{start time}

\item{te:}{end time}
}
\description{
Function to download covariates for a given area and time period
}
