% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_era5.R
\name{download_era5}
\alias{download_era5}
\title{Download ERA5 reanalysis of temperature for a given period and area
data is stored per month in a netcdf file
/!\ download quite long...}
\usage{
download_era5(user_id, years, months, latn, latx, lonn, lonx, storage_dir)
}
\arguments{
\item{user_id}{character with user Copernicus UID}

\item{years}{vector of years to download}

\item{months}{vector of characters for months to download
(e.g. c("01", "02", "03")).}

\item{latn}{minimum latitude}

\item{latx}{maximum latitude}

\item{lonn}{minimum longitude}

\item{lonx}{maximum longitude}

\item{storage_dir}{directory to save the downloaded files}
}
\description{
Download ERA5 reanalysis of temperature for a given period and area
data is stored per month in a netcdf file
/!\ download quite long...
}
