% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ghcnh.R
\name{download_ghcnh}
\alias{download_ghcnh}
\title{Download all GHCN-H stations data in area between two dates}
\usage{
download_ghcnh(ts, te, area)
}
\arguments{
\item{ts}{start date}

\item{te}{end date}

\item{area}{a sf, sfc, SpatRaster or SpatVector object}
}
\value{
a data.frame with the GHCN-H stations observations in the area
}
\description{
Download all GHCN-H stations data in area between two dates
}
\references{
Menne, Matthew J.; Noone, Simon; Casey, Nancy W.; Dunn, Robert H.; McNeill, Shelley; Kantor, Diana; Thorne, Peter W.; Orcutt, Karen; Cunningham, Sam; Risavi, Nicholas. 2023. Global Historical Climatology Network-Hourly (GHCNh). NOAA National Centers for Environmental Information. \link{accessed on 2024/06/12}
}
