% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ghcnh.R
\name{download_ghcnh_station}
\alias{download_ghcnh_station}
\title{Download GHCN-H station data of a given year}
\usage{
download_ghcnh_station(site_id, year)
}
\arguments{
\item{site_id}{the GHCN-H station ID}

\item{year}{the year of the data}
}
\value{
a data.frame with the GHCN-H station raw data
}
\description{
Download GHCN-H station data of a given year
}
\references{
Menne, Matthew J.; Noone, Simon; Casey, Nancy W.; Dunn, Robert H.; McNeill, Shelley; Kantor, Diana; Thorne, Peter W.; Orcutt, Karen; Cunningham, Sam; Risavi, Nicholas. 2023. Global Historical Climatology Network-Hourly (GHCNh). NOAA National Centers for Environmental Information. \link{accessed on 2024/06/12}
}
\author{
Eva Marques
}
