% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_covariates.R
\name{extract_building_footprint}
\alias{extract_building_footprint}
\title{Extracts building footprint at spatial points}
\usage{
extract_building_footprint(pts, bf, buf_radius = 500)
}
\arguments{
\item{pts}{a SpatVector or sf (should not be spatiotemporal)
\code{\link[=create_grids_from_cells]{create_grids_from_cells()}}).}

\item{bf}{a SpatRaster with building footprint data.}

\item{buf_radius}{a numeric with the radius of the buffer around each point.}
}
\description{
Extracts building footprint at spatial points
}
