% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_covariates.R
\name{extract_era5}
\alias{extract_era5}
\title{Extracts era5 covariates}
\usage{
extract_era5(
  pts_st,
  era5_accum,
  era5_instant,
  era5_accum_path,
  era5_instant_path,
  ts,
  te
)
}
\arguments{
\item{pts_st}{a SpatVector or sf (should be spatiotemporal)}

\item{era5_accum}{a SpatRaster with era5 accumulated covariates}

\item{era5_instant}{a SpatRaster with era5 instantaneous covariates}

\item{era5_accum_path}{character path to era5_accum netcdf to extract
some correct timestamps}

\item{era5_instant_path}{character path to era5_instant netcdf to extract
some correct timestamps}

\item{ts}{POSIXct of the start timestamp}

\item{te}{POSIXct of the end timestamp}
}
\description{
Extracts era5 covariates at spatial points
}
\author{
Eva Marques
}
