% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_covariates.R
\name{extract_evapotranspiration}
\alias{extract_evapotranspiration}
\title{Extracts evapotranspiration at spatial points}
\usage{
extract_evapotranspiration(pts, et, buf_radius = 500)
}
\arguments{
\item{pts}{a SpatVector or sf (should not be spatiotemporal)}

\item{et}{a SpatRaster with building footprint data.}

\item{buf_radius}{a numeric with the radius of the buffer around each point.}
}
\description{
Extracts evapotranspiration at spatial points
}
