% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_covariates.R
\name{extract_forest_canopy_height}
\alias{extract_forest_canopy_height}
\title{Extracts forest canopy height at spatial points}
\usage{
extract_forest_canopy_height(pts, fch, buf_radius = 500)
}
\arguments{
\item{pts}{a SpatVector or sf (should not be spatiotemporal)
\code{\link[=create_grids_from_cells]{create_grids_from_cells()}}).}

\item{fch}{a SpatRaster with forest canopy height data.}

\item{fsave}{a character with the file path to save the extracted data.}
}
\description{
Extracts forest canopy height at spatial points
}
