% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_covariates.R
\name{extract_imperviousness}
\alias{extract_imperviousness}
\title{Extracts imperviousness at spatial points}
\usage{
extract_imperviousness(pts, imp, buf_radius = 500)
}
\arguments{
\item{pts}{a SpatVector or sf (should not be spatiotemporal)
\code{\link[=create_grids_from_cells]{create_grids_from_cells()}}).}

\item{imp}{a SpatRaster with imperviousness data.}

\item{fsave}{a character with the file path to save the extracted data.}
}
\description{
Extracts imperviousness at spatial points
}
