% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_covariates.R
\name{extract_tree_canopy_cover}
\alias{extract_tree_canopy_cover}
\title{Extracts tree canopy cover at spatial points}
\usage{
extract_tree_canopy_cover(pts, tcc, buf_radius = 500)
}
\arguments{
\item{pts}{a SpatVector or sf (should not be spatiotemporal)
\code{\link[=create_grids_from_cells]{create_grids_from_cells()}}).}

\item{tcc}{a SpatRaster with tree canopy cover data.}

\item{buf_radius}{a numeric with the radius of the buffer around each point.}
}
\description{
Extracts tree canopy cover at spatial points
}
