% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_prediction_grid.R
\name{find_cells}
\alias{find_cells}
\title{Find cells id in a polygon}
\usage{
find_cells(polygon)
}
\arguments{
\item{polygon}{a terra::SpatVector polygon}
}
\description{
Find cells id in a polygon
}
