% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ghcnh.R
\name{find_ghcnh_polygon}
\alias{find_ghcnh_polygon}
\title{Find all the GHCN-H stations within a polygon}
\usage{
find_ghcnh_polygon(polygon)
}
\arguments{
\item{polygon}{a sf object with a POLYGON geometry}
}
\value{
a sf object with the GHCN-H stations within the polygon
}
\description{
Find all the GHCN-H stations within a polygon
}
\references{
Menne, Matthew J.; Noone, Simon; Casey, Nancy W.; Dunn, Robert H.; McNeill, Shelley; Kantor, Diana; Thorne, Peter W.; Orcutt, Karen; Cunningham, Sam; Risavi, Nicholas. 2023. Global Historical Climatology Network-Hourly (GHCNh). NOAA National Centers for Environmental Information. \link{accessed on 2024/06/12}
}
\author{
Eva Marques
}
