% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ghcnh.R
\name{format_area}
\alias{format_area}
\title{Tranform spatial objects to a POLYGON surrounding an area}
\usage{
format_area(x)
}
\arguments{
\item{x}{a sf, sfc, SpatRaster or SpatVector object}
}
\value{
a POLYGON object or a bbox
}
\description{
Tranform spatial objects to a POLYGON surrounding an area
}
