% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_spatiotemporal_model.R
\name{inference}
\alias{inference}
\title{Spatiotemporal model inference with INLA}
\usage{
inference(data, pred, polygon, ts, te, verbose = FALSE, debug = FALSE)
}
\arguments{
\item{data}{a dataframe with columns lon, lat, time, temp and all covariates.
Coordinates are in EPSG:4326.}

\item{pred}{a dataframe with columns lon, lat, time and all covariates.
Coordinates are in EPSG:4326.}

\item{polygon}{a terra::SpatVector with the study area}

\item{ts}{model start time}

\item{te}{model end time}

\item{verbose}{logical, print INLA output}
}
\value{
a list with the prediction and the model object
}
\description{
Perform inference of a spatiotemporal model using INLA.
}
\author{
Eva Marques
}
