% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflate.R
\name{inflate}
\alias{inflate}
\title{Inflate the prediction grid raster to a spatiotemporal datatable}
\usage{
inflate(s, yyyy, mm)
}
\arguments{
\item{s}{a terra::SpatVector of points (eg: points of a 100km * 100km grid)}

\item{yyyy}{a numeric year}

\item{mm}{a numeric month}
}
\value{
a sf object with a time column and duplicated spatial covariates
along the time dimension
}
\description{
Inflate the prediction grid raster to a spatiotemporal datatable
with dates
}
