% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Samba_functions.R
\name{KStest_Samba}
\alias{KStest_Samba}
\title{KStest_Samba}
\usage{
KStest_Samba(data, nullData, direction = 1)
}
\arguments{
\item{data}{data.frame of the results from the guide-level screen enrichment analysis.}

\item{nullData}{data.frame of randomly arranged null-genes. This is used to create the null score distribution.}

\item{direction}{integer of either 1 or -1, indicating whether to perform enrichment or depletion analysis, respectively.}
}
\value{
data.frame of the results from the gene-level screen enrichment analysis.
}
\description{
This function performs SAMBA gene-level aggregation, using the
  Kolmogorov-Smirnov test. Note that this is an internal function.
}
