% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_palette.R
\name{load_palette}
\alias{load_palette}
\title{Load palettes}
\usage{
load_palette(name)
}
\arguments{
\item{name}{character}
}
\value{
list of colors
}
\description{
Load useful homemade palettes for plots
}
\author{
Eva Marques
}
