% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_pred_vs_pro.R
\name{map_pred_with_pro}
\alias{map_pred_with_pro}
\title{Map prediction mean with professional measurements for a given time}
\usage{
map_pred_with_pro(pred, pro, ts, borders, tz)
}
\arguments{
\item{pred}{prediction dataset with pred_mean column}

\item{pro}{dataset of professional weather stations used as reference}

\item{ts}{time stamp for the map}

\item{borders}{borders shapefile (as a landmark on the map)}

\item{tz}{string of the timezone chosen to be printed with the timestamp
in the title}
}
\value{
A ggplot object displaying the residual median evolution through time
}
\description{
Map prediction mean with professional measurements for a given time
}
