% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_pred_vs_pro.R
\name{map_rmse_nlcd}
\alias{map_rmse_nlcd}
\title{Calculate and plot RMSE map with NLCD background
This function calculates the root mean square error (RMSE)
for each station in the \code{pro} dataset and plots it on a map
with National Land Cover Dataset background}
\usage{
map_rmse_nlcd(pro, nlcd)
}
\arguments{
\item{pro}{The dataset of professional weather stations used as reference}

\item{nlcd}{National Land Cover Dataset raster}
}
\value{
A ggplot object displaying the RMSE values on a map.
}
\description{
Calculate and plot RMSE map with NLCD background
This function calculates the root mean square error (RMSE)
for each station in the \code{pro} dataset and plots it on a map
with National Land Cover Dataset background
}
