% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_predictions.R
\name{map_uhi}
\alias{map_uhi}
\title{Map urban heat island}
\usage{
map_uhi(uhi, borders, ts)
}
\arguments{
\item{uhi}{urban heat island SpatRaster}

\item{borders}{vector with borders shapefile (as a landmark on the map)}

\item{ts}{POSIXct time of the map}
}
\value{
a ggplot2 of a map
}
\description{
Map urban heat island infered with the Bayesian model
}
\author{
Eva Marques
}
