% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_predictions.R
\name{map_uhi_obs}
\alias{map_uhi_obs}
\title{Map urban heat island with point observations}
\usage{
map_uhi_obs(pred_stat, obs, borders, time)
}
\arguments{
\item{pred_stat}{a dataframe of predictions}

\item{obs}{a dataframe of point observations}

\item{borders}{vector with borders shapefile (as a landmark on the map)}

\item{time}{POSIXct time of the map}
}
\value{
a ggplot2 of a map
}
\description{
Map urban heat island infered with the Bayesian model
with observations from weather stations
}
\author{
Eva Marques
}
