% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_pred_vs_pro.R
\name{median_bias_per_hour}
\alias{median_bias_per_hour}
\title{Plot median residuals per hour}
\usage{
median_bias_per_hour(pro)
}
\arguments{
\item{pro}{The dataset of professional weather stations used as reference}
}
\value{
A ggplot object displaying the residual median evolution through time
}
\description{
Plot median residuals per hour
}
