% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAMBA_Functions.R
\name{MetaAnalysisSamba}
\alias{MetaAnalysisSamba}
\title{MetaAnalysisSamba}
\usage{
MetaAnalysisSamba(data, nullData, fdr.threshold, direction = 1)
}
\arguments{
\item{data}{data.frame of the results from the guide-level screen enrichment analysis.}

\item{nullData}{data.frame of guide-level screen enrichment data for the null distribution.}

\item{fdr.threshold}{numeric between 0 and 1, indicating the FDR to use as a threshold.}

\item{direction}{integer of either 1 or -1, indicating whether to perform enrichment or depletion analysis, respectively.}
}
\value{
data.frame of the results from the gene-level screen enrichment analysis.
}
\description{
This function performs SAMBA gene-level aggregation, using the meta-analysis method. Note that this is an internal function.
}
