% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_meterological_conditions.R
\name{mytile}
\alias{mytile}
\title{Generic function to create meteorological tile plot on one month}
\usage{
mytile(df, fill)
}
\arguments{
\item{df}{dataframe. Contains the covariate to be plotted by the tile plot
and time column}

\item{fill}{character. Name of the covariate to be plotted.}
}
\value{
a ggplot2 tileplot
}
\description{
Generic function to create meteorological tile plot on one month
}
\author{
Eva Marques
}
