% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_meterological_conditions.R
\name{mytile_rain}
\alias{mytile_rain}
\title{Rain tile plot on one month}
\usage{
mytile_rain(df)
}
\arguments{
\item{df}{dataframe. Contains \code{rain} and \code{time} columns.}
}
\value{
a ggplot2 tileplot
}
\description{
Rain tile plot on one month (rain is categorical)
}
\author{
Eva Marques
}
