% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlogistic.R
\name{nonlogistic}
\alias{nonlogistic}
\title{nonlogistic}
\usage{
nonlogistic(Z, X, Dstar, weights = NULL, c_X)
}
\arguments{
\item{Z}{matrix or data frame with covariates in disease model}

\item{X}{matrix or data frame with covariates in sensitivity model. Set to NULL to fit model with no covariates in sensitivity model.}

\item{Dstar}{matrix or data frame containing observed disease status}

\item{weights}{(optional) vector of subject-specific weights used for selection bias adjustment.}

\item{c_X}{sensitivity as a function of X, P( observe disease | have disease, X )}
}
\value{
param vector with parameter estimates for disease model (intercept, logOR of Z)

variance vector of variance estimates for disease model
}
\description{
This function can be used to estimate parameters in the disease model given previously-estimated sensitivity. This function fits a logistic regression model for D given Z using a non-logistic link function for Dstar given Z and sensitivity.
}
\details{
We are interested in modeling the relationship between binary disease status and covariates Z using a logistic regression model. However, D may be misclassified, and our observed data may not well-represent the population of interest. In this setting, we estimate parameters from the disease model using the following modeling framework.

Notation:

D = binary disease status of interest

Dstar = observed binary disease status. Potentially a misclassified version of D. We assume D=0 implies Dstar=0.

S = indicator for whether patient from population of interest is included in the analytical dataset

Z = covariates in disease model of interest

W = covariates in model for patient inclusion in analytical dataset (selection model)

X = covariates in model for probability of observing disease given patient has disease (sensitivity model)

Model Structure:

Disease Model: logit(P(D=1|X)) = theta_0 + theta_Z Z

Selection Model: P(S=1|W,D)

Sensitivity Model: logit(P(Dstar=1|D=1,X)) = beta_0 + beta_X X
}
