% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsloglik_weightedvar.R
\name{obsloglik_weightedvar}
\alias{obsloglik_weightedvar}
\title{obsloglik_weightedvar}
\usage{
obsloglik_weightedvar(Dstar, X, Z, theta, beta, weights = NULL,
  getInfo = FALSE, expectedInfo = TRUE)
}
\arguments{
\item{Dstar}{Numeric vector containing observed disease status. Should be
coded as 0/1}

\item{X}{Numeric matrix with covariates in sensitivity model. Set to NULL
to fit model with no covariates in sensitivity model. 'X' should not
contain an intercept}

\item{Z}{Numeric matrix of covariates in disease model}

\item{theta}{estimated value of theta from a call to misclass_max or misclass_maxEM}

\item{beta}{estimated value of beta from a call to misclass_max or misclass_maxEM}

\item{weights}{Optional vector of subject-specific weights used for
selection bias adjustment. Default is NULL}

\item{getInfo}{Logical variable for whether the information matrix or
covariance matrix should be returned. Default is FALSE.}

\item{expectedInfo}{Logical variable for whether the observed or expected
information matrix should be used to estimate variance. Default is the
expected information matrix (TRUE).}
}
\value{
Info estimated information matrix.

variance estimated covariance matrix.
}
\description{
This function can be used to estimate the observed data information matrix
for joint estimation of parameters in the disease and sensitivity models.
}
\details{
We are interested in modeling the relationship between binary disease status
and covariates Z using a logistic regression model. However, D may be
misclassified, and our observed data may not well-represent the population
of interest. In this setting, we estimate parameters from the disease model
using the following modeling framework.
Notation:
\describe{
    \item{D}{Binary disease status of interest.}
    \item{Dstar}{Observed binary disease status. Potentially a misclassified
                 version of D. We assume D = 0 implies Dstar = 0.}
    \item{S}{Indicator for whether patient from population of interest is
             included in the analytical dataset.}
    \item{Z}{Covariates in disease model of interest.}
    \item{W}{Covariates in model for patient inclusion in analytical dataset
             (selection model).}
    \item{X}{Covariates in model for probability of observing disease given
             patient has disease (sensitivity model).}
}
Model Structure:
\describe{
\item{Disease Model}{\deqn{logit(P(D=1|X)) = theta_0 + theta_Z Z}}
\item{Selection Model}{\deqn{P(S=1|W,D)}}
\item{Sensitivity Model}{\deqn{logit(P(D^*=1|D=1,X)) = beta_0 + beta_X X}}
}
}
\note{
This function assumes that the sensitivity model intercept was
estimated rather than fixed. If it was fixed, this will produce variances
that are too big. When the intercept is fixed, (1) return the information
matrix, (2) remove the row/column corresponding to the fixed parameter,
and (3) invert the information matrix to get the fixed-parameter covariance
matrix
}
