% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wind_rose.R
\name{plot_windrose}
\alias{plot_windrose}
\title{Plot wind rose.}
\usage{
plot_windrose(
  data,
  spd,
  dir,
  spdres = 2,
  dirres = 30,
  spdmin = 0,
  spdmax = 20,
  spdseq = NULL,
  palette = "YlGnBu",
  countmax = NA
)
}
\arguments{
\item{data}{data.frame.}

\item{spd}{vector. Wind speed.}

\item{dir}{vector. Wind direction.}

\item{spdres}{integer. Resolution of wind speed scale.}

\item{dirres}{integer. Resolution of wind direction scale.}

\item{spdmin}{integer. Minimum of wind speed.}

\item{spdmax}{integer. Maximum of wind speed.}

\item{palette}{character. Palette for the wind speed.}

\item{countmax}{integer. y-axis adjustments.}
}
\value{
a ggplot2 of the form of a wind rose
}
\description{
Plot wind rose.
}
\references{
Code from \link{https://stackoverflow.com/questions/17266780/wind-rose-with-ggplot-r}
}
