% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_bhm_materials.R
\name{prepare_bhm_materials}
\alias{prepare_bhm_materials}
\title{Process elevation raster}
\usage{
prepare_bhm_materials(input, era5_accum_path, era5_instant_path)
}
\description{
Process elevation raster
}
