% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAMBA_Functions.R
\name{Preprocess_Samba}
\alias{Preprocess_Samba}
\title{Preprocess_Samba}
\usage{
Preprocess_Samba(data, design)
}
\arguments{
\item{data}{data.frame with read counts. Must have "sgRNA" and "Gene" columns, followed by columns with raw sample counts.}

\item{design}{design matrix for the samples in "data". Note: The order of the rows must match that of the data read-counts.}
}
\value{
DGEList object of the edgeR package.
}
\description{
This function preprocesses read count data from CRISPR screen data. The structure of the
input data is assumed to have columns labeled "sgRNA" and "Gene", followed by read counts
from each sample in appropriately labeled columns.
}
