% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_covariates.R
\name{process_bf}
\alias{process_bf}
\title{Process building footprint raster}
\usage{
process_bf(bf, polygon)
}
\arguments{
\item{bf}{RasterLayer: building footprint raster}

\item{polygon}{SpatVector polygon to crop raster}
}
\value{
RasterLayer: processed building footprint raster
}
\description{
Process building footprint raster: select extent,
and set NA values
}
