% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_covariates.R
\name{process_et}
\alias{process_et}
\title{Process evapotranspiration raster}
\usage{
process_et(modis_et, polygon)
}
\arguments{
\item{modis_et}{RasterLayer: modis raster with evapotranspiration layer}

\item{polygon}{SpatVector polygon to crop raster}
}
\value{
RasterLayer: processed evapotranspiration raster
}
\description{
Process evapotranspiration raster: select extent,
and set NA values
}
