% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_covariates.R
\name{process_fch}
\alias{process_fch}
\title{Process forest canopy height raster}
\usage{
process_fch(fch, polygon)
}
\arguments{
\item{fch}{RasterLayer: forest canopy height raster}

\item{polygon}{SpatVector polygon to crop raster}
}
\value{
RasterLayer: processed forest canopy height raster
}
\description{
Process forest canopy height raster: select extent,
and set NA values
}
