% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_covariates.R
\name{process_imp}
\alias{process_imp}
\title{Process imperviousness raster}
\usage{
process_imp(imp, polygon)
}
\arguments{
\item{imp}{RasterLayer: imperviousness raster}

\item{polygon}{SpatVector polygon to crop raster}
}
\value{
RasterLayer: processed imperviousness raster
}
\description{
Process imperviousness raster: select extent,
and set NA values
}
