% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_covariates.R
\name{process_lcz}
\alias{process_lcz}
\title{Process local climate zone raster}
\usage{
process_lcz(lcz, polygon)
}
\arguments{
\item{lcz}{RasterLayer: local climate zone raster}

\item{polygon}{SpatVector polygon to crop raster}
}
\value{
RasterLayer: processed local climate zone raster
}
\description{
Process local climate zone raster: select extent,
and set NA values
}
