% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Samba_functions.R
\name{RandomIndexGenerator}
\alias{RandomIndexGenerator}
\title{RandomIndexGenerator}
\usage{
RandomIndexGenerator(sgPerGene, nTotalGuides)
}
\arguments{
\item{sgPerGene}{integer vector, representing the number of guides for each gene.}

\item{nTotalGuides}{integer, indicating the total number of guides in the filtered library.}
}
\value{
list of integer vectors, each of random numbers from which to construct a null data distribution.
}
\description{
This function creates a list of reproducible, random integers from which to create a null distribution.
}
