% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAMBA-package.R
\docType{package}
\name{SAMBA-package}
\alias{SAMBA-package}
\title{SAMBA: CRISPR Screen analysis with moderated Bayesian statistics and aggregated gene scoring}
\description{
This analysis algorithm employs generalized linear models (GLMs) to fit CRISPR screen data, and it analyzes guide-level enrichment/depletion using moderated Bayesian statistics and quasi-likelihood F tests. Gene-level scores are then calculated and compared with a null distribution to calculate enrichment and depletion for the screen.
}
\author{
\strong{Maintainer}: Paul Renauer \email{paul.renauer@yale.edu} (\href{https://orcid.org/0000-0002-8747-0552}{ORCID})

}
\keyword{internal}
