% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samba.df}
\alias{samba.df}
\title{Synthetic example data for SAMBA adapted from the vignette}
\format{A synthetic data.frame with 4999 observations on 5 variables:
\describe{
  \item{X}{Covariate for sensitivity model.}
  \item{Z}{Covariate for disease model.}
  \item{W}{Selection Covariate}
  \item{D}{True disease status.}
  \item{Dstar}{Observed disease status.}
}}
\usage{
samba.df
}
\description{
'samba.df' is the sampled data from the entire population
}
\keyword{datasets}
