% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_terra.R
\name{sf_as_spatraster}
\alias{sf_as_spatraster}
\title{Convert an \code{sftime} to a \code{SpatRaster}}
\usage{
sf_as_spatraster(x, varname, nx, ny)
}
\arguments{
\item{x}{an \code{sftime} object}

\item{varname}{variable to rasterize}
}
\value{
a \code{SpatRaster} object
}
\description{
Convert an \code{sftime} object to a \code{SpatRaster} object. Returns a
\code{SpatRatser} with one layer for each time step in \code{x}.
}
\note{
Running \code{sftime_as_spatraster} can take a long time if \code{x} is not
spatially structured.
}
\seealso{
\link[terra:rast]{terra::rast}
}
\author{
Eva Marques
}
\keyword{spacetime}
