% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_post_info.R
\name{store_post_info}
\alias{store_post_info}
\title{Store marginal posterior statistics
Store marginal posterior statistics}
\usage{
store_post_info(mod, info)
}
\arguments{
\item{mod}{Bayesian model output}

\item{info}{list containing all Bayesian model information}
}
\value{
info list with appended marginal posterior statistics
}
\description{
Store marginal posterior statistics
Store marginal posterior statistics
}
\author{
Eva Marques
}
