% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_pred_vs_pro.R
\name{timeseries_res}
\alias{timeseries_res}
\title{Plot residual timeseries}
\usage{
timeseries_res(pro)
}
\arguments{
\item{pro}{professional weather station dataset}
}
\value{
a plot of the residual timeseries
}
\description{
Plot residual timeseries
}
\author{
Eva Marques
}
