% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_pred_vs_pro.R
\name{timeseries_temp}
\alias{timeseries_temp}
\title{Plot temperature timeseries}
\usage{
timeseries_temp(pred, pro, ts, te)
}
\arguments{
\item{pred}{prediction dataset}

\item{pro}{professional weather station dataset}

\item{ts}{POSIXct of timeserie starting date}

\item{te}{POSIXct of timeserie ending date}
}
\value{
a plot of the timeseries
}
\description{
Plot predicted and observed (by professional network) temperature timeseries
}
\author{
Eva Marques
}
