% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approxdist.R
\name{approxdist}
\alias{approxdist}
\title{Estimate parameters in the disease model approximating the observed data
distribution}
\usage{
approxdist(Dstar, Z, c_marg, weights = NULL)
}
\arguments{
\item{Dstar}{Numeric vector containing observed disease status. Should be
coded as 0/1}

\item{Z}{Numeric matrix of covariates in disease model}

\item{c_marg}{marginal sensitivity, P(D* = 1 | D = 1, S = 1)}

\item{weights}{Optional numeric vector of patient-specific weights used for
selection bias adjustment. Default is NULL}
}
\value{
a list with two elements: (1) 'param', a vector with parameter
    estimates for disease model (logOR of Z), and (2) 'variance', a vector of
    variance estimates for disease model parameters. Results do not include
    intercept.
}
\description{
\code{approxdist} estimates parameters in the disease model given
a previously-estimated marginal sensitivity. This estimation is based on
approximating the distribution of D* given Z.
}
\details{
We are interested in modeling the relationship between binary disease status
and covariates Z using a logistic regression model. However, D may be
misclassified, and our observed data may not well-represent the population
of interest. In this setting, we estimate parameters from the disease model
using the following modeling framework.

Notation:
\describe{
    \item{D}{Binary disease status of interest.}
    \item{D*}{Observed binary disease status. Potentially a misclassified
                 version of D. We assume D = 0 implies D* = 0.}
    \item{S}{Indicator for whether patient from population of interest is
             included in the analytical dataset.}
    \item{Z}{Covariates in disease model of interest.}
    \item{W}{Covariates in model for patient inclusion in analytical dataset
             (selection model).}
    \item{X}{Covariates in model for probability of observing disease given
             patient has disease (sensitivity model).}
}
Model Structure:
\describe{
\item{Disease Model}{\deqn{logit(P(D=1|X)) = theta_0 + theta_Z Z}}
\item{Selection Model}{\deqn{P(S=1|W,D)}}
\item{Sensitivity Model}{\deqn{logit(P(D* = 1| D = 1, S = 1, X)) = beta_0 + beta_X X}}
}
}
