% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_covariates.R
\name{extract_elevation}
\alias{extract_elevation}
\title{Extracts elevation at spatial points}
\usage{
extract_elevation(pts, elev, buf_radius = 500)
}
\arguments{
\item{pts}{a SpatVector or sf (should not be spatiotemporal)
\code{\link[=create_grids_from_cells]{create_grids_from_cells()}}).}

\item{elev}{a SpatRaster with elevation data.}

\item{buf_radius}{a numeric with the radius of the buffer around each point.}
}
\description{
Extracts elevation at spatial points
}
