% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_covariates.R
\name{extract_era5}
\alias{extract_era5}
\title{Extracts era5 at spatiotemporal points}
\usage{
extract_era5(pts, era5)
}
\arguments{
\item{pts}{a SpatVector with "time" column (spatiotemporal points).
Should be a 100km*100km hourly grid with 1km resolution}

\item{era5}{a SpatRaster with era5 data. Layers correspond to time (hours).}
}
\description{
Extracts era5 at spatiotemporal points
}
