% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_covariates.R
\name{extract_local_climate_zone}
\alias{extract_local_climate_zone}
\title{Extracts local climate zone at spatial points}
\usage{
extract_local_climate_zone(pts, lcz, buf_radius = 500)
}
\arguments{
\item{pts}{a SpatVector or sf (should not be spatiotemporal)}

\item{lcz}{a SpatRaster with local climate zone data.}

\item{buf_radius}{a numeric with the radius of the buffer around each point.}
}
\description{
Extracts local climate zone at spatial points
}
