% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Samba_functions.R
\name{FilterCountData}
\alias{FilterCountData}
\title{FilterCountData}
\usage{
FilterCountData(dge, group = NULL, min.guides, verbose = T)
}
\arguments{
\item{dge}{DGEList object with raw count data.}

\item{group}{factor, indicating the grouping of all samples. Default is NULL.}

\item{min.guides}{numeric, indicating the minimum number of counts for each guide across all samples. Default is 0.}
}
\value{
character vector of the guides that passed filtering.
}
\description{
This function filters DGEList of raw count data, based on a minimum number of total counts.
}
