% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflate.R
\name{inflate}
\alias{inflate}
\title{Inflate the prediction grid raster to a spatiotemporal SpatVector}
\usage{
inflate(s, ts, te)
}
\arguments{
\item{s}{a terra::SpatVector of points (eg: points of a 100km * 100km grid)}

\item{ts}{starting time}

\item{te}{ending time}
}
\value{
a spatVector object with a time column
and duplicated spatial covariates
along the time dimension
}
\description{
Inflate the prediction grid raster
to a spatiotemporal SpatVector
with datetime in "time" column
}
\author{
Eva Marques
}
