% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_outputs.R
\name{map_pred_mean}
\alias{map_pred_mean}
\title{Plot maps of the Bayesian model prediction mean}
\usage{
map_pred_mean(pred, borders, tz = "America/New_York")
}
\arguments{
\item{pred}{data.frame. Contains \code{pred_mean}, \code{lat}, \code{lon}, \code{time} columns.}

\item{borders}{\code{terra::SpatVector}. Borders used as a landmark on the map.}

\item{tz}{character. Timezone to write on the map.}
}
\value{
a list of ggplot2 objects
}
\description{
Plot map of the Bayesian model prediction mean
}
\author{
Eva Marques
}
