% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_outputs.R
\name{map_pred_mean}
\alias{map_pred_mean}
\title{Plot model outputs}
\usage{
map_pred_mean(pred, borders, tz = "America/New_York")
}
\arguments{
\item{pred}{a data.frame with prediction mean, lat, lon, time}

\item{borders}{a terra::SpatVector with territory borders}

\item{tz}{timezone character}
}
\value{
a list of ggplot2 objects
}
\description{
Plot model prediction mean
}
