% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_pred_vs_pro.R
\name{map_rmse}
\alias{map_rmse}
\title{Calculate and plot RMSE map}
\usage{
map_rmse(pro, imp)
}
\arguments{
\item{pro}{The datatable of professional weather stations used as reference}

\item{imp}{Imperviousness raster}
}
\value{
A ggplot object displaying the RMSE values on a map.
}
\description{
This function calculates the root mean square error (RMSE)
for each station in the \code{pro} dataset and plots it on a map.
}
