% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlogistic.R
\name{nonlogistic}
\alias{nonlogistic}
\title{Estimate parameters in the disease model given a previously-estimated
sensitivity.}
\usage{
nonlogistic(Dstar, Z, c_X, weights = NULL)
}
\arguments{
\item{Dstar}{Numeric vector containing observed disease status. Should be
coded as 0/1}

\item{Z}{numeric matrix of covariates in disease model}

\item{c_X}{sensitivity as a function of X, P(D^* | D, X)}

\item{weights}{(optional) vector of subject-specific weights used for
selection bias adjustment.}
}
\value{
A list with two elements: 'param', a vector with parameter estimates
    for disease model(intercept, logOR of Z), and 'variance', a vector of
    variance estimates for disease model
}
\description{
\code{nonlogistic} fits a logistic regression model for D given Z using a
non-logistic link function for Dstar given Z and sensitivity.
}
\details{
We are interested in modeling the relationship between binary disease status
and covariates Z using a logistic regression model. However, D may be
misclassified, and our observed data may not well-represent the population
of interest. In this setting, we estimate parameters from the disease model
using the following modeling framework.

Notation:
\describe{
    \item{D}{Binary disease status of interest.}
    \item{Dstar}{Observed binary disease status. Potentially a misclassified
                 version of D. We assume D = 0 implies Dstar = 0.}
    \item{S}{Indicator for whether patient from population of interest is
             included in the analytical dataset.}
    \item{Z}{Covariates in disease model of interest.}
    \item{W}{Covariates in model for patient inclusion in analytical dataset
             (selection model).}
    \item{X}{Covariates in model for probability of observing disease given
             patient has disease (sensitivity model).}
}
Model Structure:
\describe{
\item{Disease Model}{\deqn{logit(P(D=1|X)) = theta_0 + theta_Z Z}}
\item{Selection Model}{\deqn{P(S=1|W,D)}}
\item{Sensitivity Model}{\deqn{logit(P(D^*=1|D=1,X)) = beta_0 + beta_X X}}
}
}
