% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsloglik.R
\name{obsloglik}
\alias{obsloglik}
\title{Calculate Observed Profile Log likelihood}
\usage{
obsloglik(Dstar, Z, X, param_current, weights = NULL,
  beta0_fixed = NULL, itnmax = 5000)
}
\arguments{
\item{Dstar}{matrix or data frame containing observed disease status}

\item{Z}{Numeric matrix with covariates in disease model}

\item{X}{Numeric Matrix with covariates in sensitivity model. Set to
NULL to fit model with no covariates in sensitivity model}

\item{param_current}{vector of starting values for theta and beta (theta, beta).
Theta is the parameter of the disease model, and beta is the parameter
of the sensitivity model}

\item{weights}{(optional) vector of subject-specific weights used for
selection bias adjustment}

\item{beta0_fixed}{(optional) vector of values of sensitivity model intercept
to profile over. If a single value, corresponds to fixing intercept at
specified value}

\item{itnmax}{Maximum number of iterations to run \code{optimx}}
}
\value{
param vector with parameter estimates organized as (theta, beta)

param_save matrix containing estimated parameter values corresponding
    to different values of beta0_fixed.

loglik_save vector of log-likelihood values corresponding to
   different values of beta0_fixed.
}
\description{
\code{obsloglik} jointly estimates the disease model and sensitivity
model parameters using profile likelihood methods.
}
\details{
We are interested in modeling the relationship between binary disease status
and covariates Z using a logistic regression model. However, D may be
misclassified, and our observed data may not well-represent the population of
interest. In this setting, we estimate parameters from the disease model
using the following modeling framework.
Notation:
\describe{
    \item{D}{Binary disease status of interest.}
    \item{Dstar}{Observed binary disease status. Potentially a misclassified
                 version of D. We assume D = 0 implies Dstar = 0.}
    \item{S}{Indicator for whether patient from population of interest is
             included in the analytical dataset.}
    \item{Z}{Covariates in disease model of interest.}
    \item{W}{Covariates in model for patient inclusion in analytical dataset
             (selection model).}
    \item{X}{Covariates in model for probability of observing disease given
             patient has disease (sensitivity model).}
}
Model Structure:
\describe{
\item{Disease Model}{\deqn{logit(P(D=1|X)) = theta_0 + theta_Z Z}}
\item{Selection Model}{\deqn{P(S=1|W,D)}}
\item{Sensitivity Model}{\deqn{logit(P(D^*=1|D=1,X)) = beta_0 + beta_X X}}
}
}
