% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsloglik.R
\name{obsloglik}
\alias{obsloglik}
\title{Calculate Observed Profile Log likelihood}
\usage{
obsloglik(Dstar, Z, X, start, beta0_fixed = NULL, weights = NULL,
  expected = TRUE, itnmax = 5000)
}
\arguments{
\item{Dstar}{Numeric vector containing observed disease status. Should be
coded as 0/1}

\item{Z}{Numeric matrix of covariates in disease model}

\item{X}{Numeric Matrix of covariates in sensitivity model. Set to
NULL to fit model with no covariates in sensitivity model}

\item{start}{Numeric vector of starting values for theta and beta (theta, beta).
Theta is the parameter of the disease model, and beta is the parameter
of the sensitivity model}

\item{beta0_fixed}{Optional numeric vector of values of sensitivity model
intercept to profile over. If a single value, corresponds to fixing
intercept at specified value. Default is NULL}

\item{weights}{Optional vector of subject-specific weights used for
selection bias adjustment. Default is NULL}

\item{expected}{Whether or not to calculate the covariance matrix via the
expected fisher information matrix. Default is TRUE}

\item{itnmax}{Maximum number of iterations to run \code{optimx}}
}
\value{
A "SAMBA.fit" object with nine elements: 'param', the maximum
likelihood estimate of the coeficients,  'var', the covariance matrix of the
final estimate, param.seq', the sequence of estimates at each value of beta0,
and 'loglik.seq', the log likelihood at each value. The rest of the elements
are Dstar', 'X', 'Z', and 'weights'.
}
\description{
\code{obsloglik} jointly estimates the disease model and sensitivity
model parameters using profile likelihood methods.
}
\details{
We are interested in modeling the relationship between binary disease status
and covariates Z using a logistic regression model. However, D may be
misclassified, and our observed data may not well-represent the population of
interest. In this setting, we estimate parameters from the disease model
using the following modeling framework.
Notation:
\describe{
    \item{D}{Binary disease status of interest.}
    \item{Dstar}{Observed binary disease status. Potentially a misclassified
                 version of D. We assume D = 0 implies Dstar = 0.}
    \item{S}{Indicator for whether patient from population of interest is
             included in the analytical dataset.}
    \item{Z}{Covariates in disease model of interest.}
    \item{W}{Covariates in model for patient inclusion in analytical dataset
             (selection model).}
    \item{X}{Covariates in model for probability of observing disease given
             patient has disease (sensitivity model).}
}
Model Structure:
\describe{
\item{Disease Model}{\deqn{logit(P(D=1|X)) = theta_0 + theta_Z Z}}
\item{Selection Model}{\deqn{P(S=1|W,D)}}
\item{Sensitivity Model}{\deqn{logit(P(D^*=1|D=1,X)) = beta_0 + beta_X X}}
}
}
