% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsloglikEM.R
\name{obsloglikEM}
\alias{obsloglikEM}
\title{obsloglikEM
This function can be used to jointly estimate disease model and sensitivity
model parameters using an expectation-maximization algorithm.}
\usage{
obsloglikEM(Dstar, Z, X, param_current, beta0_fixed = NULL,
  weights = NULL, expected = TRUE, tol = 1e-06, maxit = 50)
}
\arguments{
\item{Dstar}{matrix or data frame containing observed disease status}

\item{Z}{matrix or data frame with covariates in disease model}

\item{X}{Numeric matrix with covariates in sensitivity model. Set to NULL
to fit model with no covariates in sensitivity model. 'X' should not
contain an intercept.}

\item{param_current}{vector of starting values for theta and beta
(theta, beta). Theta is the parameter of the disease model, and beta is
the parameter of the sensitivity model.}

\item{beta0_fixed}{(optional) fixed value for the intercept of the
sensitivity model.}

\item{weights}{(optional) vector of subject-specific weights used for
selection bias adjustment.}

\item{expected}{Whether or not to calculate the covariance matrix via the
expected fisher information matrix. Default is TRUE}

\item{tol}{stop estimation when subsequent log-likelihood estimates are
within this value}

\item{maxit}{maximum number of iterations of the estimation algorithm}
}
\value{
param vector with parameter estimates organized as (theta, beta)

param.seq matrix containing estimated parameter values across
    iterations of the expectation algorithm

loglik.seq vector of log-likelihood values across iterations of the
    expectation algorithm
}
\description{
We are interested in modeling the relationship between binary disease status
and covariates Z using a logistic regression model. However, D may be
misclassified, and our observed data may not well-represent the population
of interest. In this setting, we estimate parameters from the disease model
using the following modeling framework.
Notation:
D = binary disease status of interest
Dstar = observed binary disease status. Potentially a misclassified version
of D. We assume D=0 implies Dstar=0.
S = indicator for whether patient from population of interest is included
in the analytical dataset
Z = covariates in disease model of interest
W = covariates in model for patient inclusion in analytical dataset
(selection model)
X = covariates in model for probability of observing disease given patient
has disease (sensitivity model)
Model Structure:
Disease Model: logit(P(D=1|X)) = theta_0 + theta_Z Z
Selection Model: P(S=1|W,D)
Sensitivity Model: logit(P(Dstar=1|D=1,X)) = beta_0 + beta_X X
}
