% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Samba_functions.R
\name{Preprocess_Samba}
\alias{Preprocess_Samba}
\title{Preprocess_Samba}
\usage{
Preprocess_Samba(
  data,
  design = NULL,
  screen.names = NULL,
  ctrl.names = NULL,
  hdr.gene = "Gene",
  hdr.guide = "sgRNA",
  coefficient = NULL,
  verbose = T
)
}
\arguments{
\item{data}{data.frame with read counts. Must have "sgRNA" and "Gene" columns, followed by columns with raw sample counts.}

\item{design}{design matrix for the samples in "data". Note: The order of the rows must match that of the data read-counts.}

\item{screen.names}{character vector of the screen sample names.}

\item{ctrl.names}{character vector of the control sample names.}

\item{hdr.gene}{character, name of column with gene names. Defaults to "Gene".}

\item{hdr.guide}{character, name of column with guide names. Defaults to "sgRNA".}

\item{coefficient}{character name or vector, indicating which coefficient of the linear model is tested to be equal to zero.}

\item{verbose}{logical, indicating whether to display verbose output information.}

\item{contrast}{contrast matrix, indicating the contrast(s) of the linear model to be tested as equal to zero (optional).}
}
\value{
DGEList object of the edgeR package.
}
\description{
This function preprocesses read count data from CRISPR screen data. The structure of the
input data is assumed to have columns labeled "sgRNA" and "Gene", followed by read counts
from each sample in appropriately labeled columns.
}
