% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_covariates.R
\name{process_elev}
\alias{process_elev}
\title{Process elevation raster}
\usage{
process_elev(elev, polygon)
}
\arguments{
\item{elev}{RasterLayer: elevation raster}

\item{polygon}{SpatVector polygon to crop raster}
}
\value{
RasterLayer: processed elevation raster
}
\description{
Process elevation raster: select extent, set NA values and
calculate slope, aspect, flow direction
}
