% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize_pred.R
\name{rasterize_pred}
\alias{rasterize_pred}
\title{Convert prediction data frame to \code{terra::SpatRaster}}
\usage{
rasterize_pred(pred, varname = "pred_mean", existing_raster = NULL)
}
\arguments{
\item{pred}{data.frame. Output of SAMBA model.}

\item{varname}{character. Variable to rasterize}

\item{existing_raster}{\code{terra::SpatRaster}. Used as a template for
the layout.}
}
\value{
a \code{SpatRaster} object
}
\description{
Convert prediction data frame to \code{terra::SpatRaster}
}
\author{
Eva Marques
}
