% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samba.R
\name{samba}
\alias{samba}
\title{Run the whole samba analysis.}
\usage{
samba(
  ts,
  te,
  area_shp,
  cws_raw,
  fch_path,
  elev_path,
  imp_path,
  era5_accum_path,
  era5_instant_path
)
}
\arguments{
\item{ts}{POSIXct. Analysis starting time.}

\item{te}{POSIXct. Analysis ending time.}

\item{area_shp}{SpatVector. Area delimitation shapefile.}

\item{cws_raw}{data.table. Dataset of personal weather stations.}

\item{fch_path}{character. Path to forest canopy height raster.}

\item{elev_path}{character. Path to elevation raster.}

\item{imp_path}{character. Path to imperviousness raster.}

\item{era5_accum_path}{character. Path to era5 accumulated netcdf.}

\item{era5_instant_path}{character. Path to era5 instantaneous netcdf.}
}
\value{
a list with samba pipeline outputs.
}
\description{
Run the whole samba analysis.
}
\author{
Eva Marques
}
