% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Samba_functions.R
\name{Samba}
\alias{Samba}
\title{Samba}
\usage{
Samba(
  data,
  design = NULL,
  screen.names = NULL,
  ctrl.names = NULL,
  coefficient = NULL,
  contrast = NULL,
  control.gene = NULL,
  score.method = "GeneScore",
  test.method = "QLF",
  GuideMap = NULL,
  file.prefix = NULL,
  verbose = T
)
}
\arguments{
\item{data}{data.frame with read counts. Must have "sgRNA" and "Gene" columns, followed by columns with raw sample counts.}

\item{design}{design matrix for the samples in "data". Note: The order of the rows must match that of the data read-counts.}

\item{screen.names}{character vector of the screen sample names.}

\item{ctrl.names}{character vector of the control sample names.}

\item{coefficient}{character name or vector, indicating which coefficient of the linear model is tested to be equal to zero.}

\item{contrast}{contrast matrix, indicating the contrast(s) of the linear model to be tested as equal to zero (optional).}

\item{control.gene}{character, gene label of the internal control guides (e.g. "NTC").}

\item{score.method}{character, indicating whether to use the 'KS' (Kolmogorov-Smirnov), 'MetaAnalysis', or 'GeneScore' method to perform gene-level analysis.}

\item{test.method}{character, indicating whether to use 'QLF' or 'LRT' tests.}

\item{GuideMap}{data.frame, maps the guides of the library to their respective genes (optional). Must have "sgRNA" and "Gene" columns.}

\item{file.prefix}{character, prefix for writing output files.}

\item{verbose}{logical, indicating whether to display verbose output information.}
}
\value{
list, contains the guide-level result data.frame and the gene-level result data.frame.
}
\description{
This is an all-in-one function to preprocesses and analyze read count data from a CRISPR
screen. The structure of the input data is assumed to have columns labeled "sgRNA" and
"Gene", followed by read counts from each sample in appropriately labeled columns.
}
