% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{sensitivity}
\alias{sensitivity}
\title{sensitivity}
\usage{
sensitivity(Dstar, X, prev, r = NULL, weights = NULL)
}
\arguments{
\item{Dstar}{Numeric vector containing observed disease status. Should be
coded as 0/1}

\item{X}{Numeric matrix with covariates in sensitivity model. Set to NULL
to fit model with no covariates in sensitivity model. 'X' should not
contain an intercept.}

\item{prev}{Disease prevalence in the population or subject-specific
\eqn{P(D|X)} in population}

\item{r}{Optional marginal sampling ratio, \eqn{P(S| D) / P(S| !D)}. Only
one of 'r' and 'weights' can be specified. Default is `NULL`}

\item{weights}{Optional vector of subject-specific weights used for
selection bias adjustment. Only one of r and weights can be specified.
Default is `NULL`}
}
\value{
A list with two elements: `c_marg`, marginal sensitivity estimate
\eqn{P(D^*| D)} and `c_X`, sensitivity as a function of X, \eqn{P(D^*| D, X)}
}
\description{
Estimate the marginal sensitivity and sensitivity as a function of
covariates X for a misclassified binary outcome.
}
\details{
We are interested in modeling the relationship between binary disease status
and covariates \eqn{Z} using a logistic regression model. However, \eqn{D}
may be misclassified, and our observed data may not well-represent the
population of interest. In this setting, we estimate parameters from the
disease model using the following modeling framework.

Notation:
\describe{
    \item{D}{Binary disease status of interest.}
    \item{Dstar}{Observed binary disease status. Potentially a misclassified
                 version of D. We assume D = 0 implies Dstar = 0.}
    \item{S}{Indicator for whether patient from population of interest is
             included in the analytical dataset.}
    \item{Z}{Covariates in disease model of interest.}
    \item{W}{Covariates in model for patient inclusion in analytical dataset
             (selection model).}
    \item{X}{Covariates in model for probability of observing disease given
             patient has disease (sensitivity model).}
}
Model Structure:
\describe{
\item{Disease Model}{\deqn{logit(P(D=1|X)) = theta_0 + theta_Z Z}}
\item{Selection Model}{\deqn{P(S=1|W,D)}}
\item{Sensitivity Model}{\deqn{logit(P(D^*=1|D=1,X)) = beta_0 + beta_X X}}
}
}
