% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize_pred.R
\name{sf_as_spatraster}
\alias{sf_as_spatraster}
\title{Convert an \code{sf} to a \code{SpatRaster}}
\usage{
sf_as_spatraster(x, varname, nx, ny)
}
\arguments{
\item{x}{\code{st::sf}.}

\item{varname}{character. Variable to rasterize.}

\item{nx}{integer. Number of cells in the x direction.}

\item{ny}{integer. Number of cells in the y direction.}
}
\value{
a \code{SpatRaster} object
}
\description{
Convert an \code{sf} object to a \code{SpatRaster} object. Returns a
\code{SpatRatser} with one layer for each time step in \code{x}.
}
\author{
Eva Marques
}
