% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_terra.R
\name{sf_as_spatraster}
\alias{sf_as_spatraster}
\title{Convert an \code{sf} to a \code{SpatRaster}}
\usage{
sf_as_spatraster(x, varname, nx, ny)
}
\arguments{
\item{x}{an \code{sf} object}

\item{varname}{variable to rasterize}

\item{nx}{number of cells in the x direction}

\item{ny}{number of cells in the y direction}
}
\value{
a \code{SpatRaster} object
}
\description{
Convert an \code{sf} object to a \code{SpatRaster} object. Returns a
\code{SpatRatser} with one layer for each time step in \code{x}.
}
\author{
Eva Marques
}
